// src/actions/chat/action-type.ts
import {ICustomFile} from "../../types/app";

export enum CHAT_ACTIONS {
	ACTION_CHAT_ROOM_REQUEST = "ACTION_CHAT_ROOM_REQUEST",
	ACTION_CHAT_GET_ROOM = "ACTION_CHAT_GET_ROOM",
	ACTION_CHAT_GET_ROOM_SUCCESS = "ACTION_CHAT_GET_ROOM_SUCCESS",
	ACTION_CHAT_GET_ROOM_MESSAGES = "ACTION_CHAT_GET_ROOM_MESSAGES",
	ACTION_CHAT_GET_ROOM_MESSAGES_SUCCESS = "ACTION_CHAT_GET_ROOM_MESSAGES_SUCCESS",
	ACTION_CHAT_GET_ROOMS = "ACTION_CHAT_GET_ROOMS",
	ACTION_CHAT_GET_ROOMS_SUCCESS = "ACTION_CHAT_GET_ROOMS_SUCCESS",
	ACTION_CHAT_SEND_MESSAGE = "ACTION_CHAT_SEND_MESSAGE",
	ACTION_CHAT_SEND_MESSAGE_SUCCESS = "ACTION_CHAT_SEND_MESSAGE_SUCCESS",
	ACTION_CHAT_SEARCH_ROOMS = "ACTION_CHAT_SEARCH_ROOMS",
	ACTION_CHAT_SEARCH_ROOMS_SUCCESS = "ACTION_CHAT_SEARCH_ROOMS_SUCCESS",
	ACTION_CHAT_SEARCH_USERS = "ACTION_CHAT_SEARCH_USERS",
	ACTION_CHAT_SEARCH_USERS_SUCCESS = "ACTION_CHAT_SEARCH_USERS_SUCCESS",
	ACTION_CHAT_GET_CONTACTS = "ACTION_CHAT_GET_CONTACTS",
	ACTION_CHAT_GET_CONTACTS_SUCCESS = "ACTION_CHAT_GET_CONTACTS_SUCCESS",
	ACTION_CHAT_CREATE_GROUP = "ACTION_CREATE_GROUP",
	ACTION_CHAT_CREATE_GROUP_SUCCESS = "ACTION_CREATE_GROUP_SUCCESS",
	ACTION_CHAT_CLEAR_ROOM_MESSAGES = "ACTION_CHAT_CLEAR_ROOM_MESSAGES",
	ACTION_CHAT_CLEAR_FOUND_ROOMS = "ACTION_CHAT_CLEAR_FOUND_ROOMS",
	ACTION_CHAT_CLEAR_FOUND_USERS = "ACTION_CHAT_CLEAR_FOUND_USERS",
	ACTION_CHAT_CLEAR_STORE = "ACTION_CHAT_CLEAR_STORE",
	ACTION_CHAT_ERROR = "ACTION_CHAT_ERROR",
}

export interface ISendMessageData {
	body: string;
	images: ICustomFile[]
}

export interface ISendMessageTargetProperty {
	type: 'private' | 'public';
	roomId?: string | undefined;
	recipientId?: string | undefined;
}